+++ docs/changelog.rst
@@ -10,7 +10,8 @@ Added
 ~~~~~
 
 - Support for OpenAPI 3.0 server variables (base_path). `#40`_
-- Support for `format: byte`. `#254`_
+- Support for ``format: byte``. `#254`_
+- Response schema conformance check. `#256`_
 
 `0.14.0`_ - 2019-11-09
 ----------------------
@@ -384,6 +385,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#256: https://github.com/kiwicom/schemathesis/issues/256
 .. _#254: https://github.com/kiwicom/schemathesis/issues/254
 .. _#248: https://github.com/kiwicom/schemathesis/issues/248
 .. _#246: https://github.com/kiwicom/schemathesis/issues/246