+++ docs/changelog.rst
@@ -13,6 +13,11 @@ Added
 - Support for ``format: byte``. `#254`_
 - Response schema conformance check in CLI / Runner. `#256`_
 
+Changed
+~~~~~~~
+
+- Verbose messages are displayed in the CLI on failed checks. `#261`_
+
 `0.14.0`_ - 2019-11-09
 ----------------------
 
@@ -385,6 +390,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#261: https://github.com/kiwicom/schemathesis/issues/261
 .. _#256: https://github.com/kiwicom/schemathesis/issues/256
 .. _#254: https://github.com/kiwicom/schemathesis/issues/254
 .. _#248: https://github.com/kiwicom/schemathesis/issues/248