+++ docs/changelog.rst
@@ -383,7 +383,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.14.0...HEAD
+.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.15.0...HEAD
+.. _0.15.0: https://github.com/kiwicom/schemathesis/compare/v0.14.0...v0.15.0
 .. _0.14.0: https://github.com/kiwicom/schemathesis/compare/v0.13.2...v0.14.0
 .. _0.13.2: https://github.com/kiwicom/schemathesis/compare/v0.13.1...v0.13.2
 .. _0.13.1: https://github.com/kiwicom/schemathesis/compare/v0.13.0...v0.13.1